+++ Drivers/template/SQLExecute.c
@@ -10,6 +10,7 @@
  *
  **********************************************************************/
 
+#include <config.h>
 #include "driver.h"
 
 SQLRETURN SQLExecute( SQLHSTMT  hDrvStmt )
@@ -24,7 +25,7 @@ SQLRETURN SQLExecute( SQLHSTMT  hDrvStmt )
     if( NULL == hStmt )
         return SQL_INVALID_HANDLE;
 
-	sprintf((char*) hStmt->szSqlMsg, "hStmt = $%08lX", hStmt );
+	sprintf((char*) hStmt->szSqlMsg, "hStmt = $%08lX", (long)hStmt );
     logPushMsg( hStmt->hLog, __FILE__, __FILE__, __LINE__, LOG_WARNING, LOG_WARNING,(char*) hStmt->szSqlMsg );
 
     if( hStmt->pszQuery == NULL )
