+++ Drivers/template/SQLFreeConnect.c
@@ -13,6 +13,7 @@
  *
  **********************************************************************/
 
+#include <config.h>
 #include "driver.h"
 
 SQLRETURN _FreeConnect( SQLHDBC hDrvDbc )
@@ -24,7 +25,7 @@ SQLRETURN _FreeConnect( SQLHDBC hDrvDbc )
     if( NULL == hDbc )
         return SQL_INVALID_HANDLE;
 
-	sprintf((char*) hDbc->szSqlMsg, "hDbc = $%08lX", hDbc );
+	sprintf((char*) hDbc->szSqlMsg, "hDbc = $%08lX", (long)hDbc );
     logPushMsg( hDbc->hLog, __FILE__, __FILE__, __LINE__, LOG_WARNING, LOG_WARNING, (char*)hDbc->szSqlMsg );
 
     if( hDbc->bConnected )
