+++ Drivers/template/SQLTransact.c
@@ -10,6 +10,7 @@
  *
  ********************************************************************/
 
+#include <config.h>
 #include "driver.h"
 
 SQLRETURN SQLTransact(	SQLHENV hDrvEnv,
@@ -23,7 +24,7 @@ SQLRETURN SQLTransact(	SQLHENV hDrvEnv,
 	if ( hDbc == SQL_NULL_HDBC )
 		return SQL_INVALID_HANDLE;
 
-	sprintf((char*) hDbc->szSqlMsg, "hDbc = $%08lX", hDbc );
+	sprintf((char*) hDbc->szSqlMsg, "hDbc = $%08lX", (long)hDbc );
 	logPushMsg( hDbc->hLog, __FILE__, __FILE__, __LINE__, LOG_WARNING, LOG_WARNING,(char*) hDbc->szSqlMsg );
